join both binaries into one
