+++ src/server/instance.rs
@@ -1,8 +1,8 @@
-use crate::async_once_cell::AsyncOnceCell;
-use crate::lsp::{self, Message};
+use crate::config::Config;
+use crate::proto;
+use crate::server::async_once_cell::AsyncOnceCell;
+use crate::server::lsp::{self, Message};
 use anyhow::{bail, Context, Result};
-use ra_multiplex::config::Config;
-use ra_multiplex::proto;
 use serde_json::{Number, Value};
 use std::collections::hash_map::Entry;
 use std::collections::HashMap;
